requirements.txt
# FILE: requirements.txt

pandas
openreview-py
requests
beautifulsoup4
lxml
tqdm
PyYAML
selenium
webdriver-manager
wordcloud
matplotlib
nltk
gensim  # 强大的主题建模库
scikit-learn # 可能需要用到其中的特征提取工具

# END OF FILE: requirements.txt